+++ spec/rubocop/cop/lint/debugger_spec.rb
@@ -8,7 +8,7 @@ describe RuboCop::Cop::Lint::Debugger do
   include_examples 'debugger', 'debugger', 'debugger'
   include_examples 'debugger', 'byebug', 'byebug'
   include_examples 'debugger', 'pry binding', %w(binding.pry binding.remote_pry
-                                                 binding.pry_remote)
+                                                 binding.pry_remote Pry.rescue)
   include_examples 'debugger',
                    'capybara debug method', %w(save_and_open_page
                                                save_and_open_screenshot